+++ pkg/apis/scheduling/v2alpha2/podgroup_types.go
@@ -34,9 +34,17 @@ import (
 type PodGroupSpec struct {
 	// MinMember defines the minimal number of members to run the PodGroup;
 	// if there are not enough resources to start all required members, the scheduler will not start anyone.
+	// Mutually exclusive with MinSubGroup.
 	// +kubebuilder:validation:Nullable
 	// +kubebuilder:validation:Minimum=1
 	MinMember *int32 `json:"minMember,omitempty" protobuf:"varint,1,opt,name=minMember"`
+
+	// MinSubGroup defines the minimal number of direct child SubGroups required for this PodGroup to be schedulable.
+	// Only applicable when SubGroups are defined.
+	// Mutually exclusive with MinMember.
+	// +kubebuilder:validation:Optional
+	// +kubebuilder:validation:Minimum=1
+	MinSubGroup *int32 `json:"minSubGroup,omitempty"`
 
 	// Queue defines the queue to allocate resource for PodGroup; if queue does not exist,
 	// the PodGroup will not be scheduled.
@@ -108,9 +116,17 @@ type SubGroup struct {
 
 	// MinMember defines the minimal number of members to run this SubGroup;
 	// if there are not enough resources to start all required members, the scheduler will not start anyone.
+	// Mutually exclusive with MinSubGroup.
 	// +kubebuilder:validation:Nullable
 	// +kubebuilder:validation:Minimum=0
 	MinMember *int32 `json:"minMember,omitempty" protobuf:"varint,2,opt,name=minMember"`
+
+	// MinSubGroup defines the minimal number of direct child SubGroups required for this SubGroup to be schedulable.
+	// Only applicable when this SubGroup has child SubGroups.
+	// Mutually exclusive with MinMember.
+	// +kubebuilder:validation:Optional
+	// +kubebuilder:validation:Minimum=0
+	MinSubGroup *int32 `json:"minSubGroup,omitempty"`
 
 	// Parent is an optional attribute that specifies the name of the parent SubGroup.
 	// Must be a valid DNS label (RFC 1123).